+++ test/e2e/query_test.go
@@ -693,7 +693,6 @@ func TestSidecarStorePushdown(t *testing.T) {
 			"__name__": "my_fake_metric",
 		},
 	})
-
 }
 
 func TestSidecarQueryEvaluation(t *testing.T) {